Move Client class into baseclient.py
The goal is to provide a somewhat generic Client class that can be used
as a base for specific clients.
